% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmap.R
\name{checkGPL}
\alias{checkGPL}
\title{Check whether the input gpl in our platform list or not}
\usage{
checkGPL(GPL = NULL)
}
\arguments{
\item{GPL}{GPL(GEO platform) number, eg: GPL570}
}
\value{
returns a boolean value
}
\description{
Check whether the input gpl in our platform list or not
}
\examples{
checkGPL('GPL570')
checkGPL('GPL15314')
checkGPL('GPL10558')
}
