% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_heatmap.R
\name{deg_heatmap}
\alias{deg_heatmap}
\title{draw a heatmap for DEG result}
\usage{
deg_heatmap(deg, genes_expr, group_list, topn = 20)
}
\arguments{
\item{deg}{the result from limma.}

\item{genes_expr}{the expression matrix}

\item{group_list, }{a vector}

\item{topn}{the number of genes in heatmap, default:20}
}
\value{
a ggplot2 style figure.
}
\description{
\code{deg_heatmap} will draw a heatmap for you.
}
\examples{
attach(GSE27533)
deg_heatmap(DEG,genes_expr,group_list)
}
