% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_volcano.R
\name{deg_volcano}
\alias{deg_volcano}
\title{draw a volcano for DEG result}
\usage{
deg_volcano(need_deg, style = 1, p_thred = 0.05, logFC_thred = 1)
}
\arguments{
\item{need_deg}{should be 3 columns : gene, logFC, p.value(or p.adjust}

\item{style}{you can try 1 or 2, default: 1}

\item{p_thred}{default:0.05}

\item{logFC_thred}{default:1}
}
\value{
a ggplot2 style figure.
}
\description{
\code{deg_volcano} will draw a volcano for you.
}
\examples{
deg=GSE27533$DEG
need_deg=data.frame(symbols=rownames(deg), logFC=deg$logFC, p=deg$P.Value)
deg_volcano(need_deg,2)
\donttest{
deg_volcano(need_deg,1)
}
}
