% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIDs.R
\name{getAnno}
\alias{getAnno}
\title{Get Probe Annotation}
\usage{
getAnno(gplnum, source = "pipe", biotype = "protein_coding")
}
\arguments{
\item{gplnum}{GPL(GEO platform) number, eg: GPL570}

\item{source}{source of probe anntation stored, one of "pipe", "bioc", "soft", "all"}

\item{biotype}{GENCODE biotypes,eg:"protein_coding","non_coding","pseudogene" ...}

\item{lncRNA}{default is FALSE,lncRNA microarrays only have pipe annotation, if choose lncRNA microarray, the parameter must set as TRUE}
}
\value{
probe annotaions
}
\description{
\code{getAnno} returns annotations for target gpl
}
\examples{
gplnum <- "GPL10332"
source <- "pipe"
biotype <- "protein_coding"
humanAnno <- getAnno(gplnum, source, biotype)
head(humanAnno)

}
