% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIDs.R
\name{getGPLlist}
\alias{getGPLlist}
\title{Get GPL list}
\usage{
getGPLlist()
}
\arguments{
\item{gplnum}{GPL(GEO platform) number, eg: GPL570}
}
\value{
a dataframe which contains 146 GEO platform information .
}
\description{
\code{getGPLlist} returns a GPL number checklist stored in package
}
