% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGPLsoft.R
\name{getGPLsoft}
\alias{getGPLsoft}
\title{Download probe annotations from GEO}
\usage{
getGPLsoft(GPL, destDir = getwd())
}
\arguments{
\item{GPL}{GPL(GEO platform) number, eg: GPL570}

\item{destDir}{The destination directory for any downloads. Defaults is the currently used directory . You may want to specify a different directory if you want to save the file for later use.}
}
\value{
{getGPLsoft} will download probe annotations from GEO ,then the output will be a dataframe
}
\description{
\code{getGPLsoft} download probe annotations from GEO database for the input GPL Accession Number
}
\examples{
getGPLsoft("GPL13607", destdir=getwd())

}
