% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE27533}
\alias{GSE27533}
\title{An example dataset}
\format{
A list with 6 elements:
\describe{
  \item{genes_expr}{genes_expr, genes_expr}
  \item{DEG}{DEG, DEG}
  ...
}
}
\usage{
GSE27533
}
\description{
A dataset containing genes_expr, group_list, DEG
}
\keyword{datasets}
