% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE95166}
\alias{GSE95166}
\title{An example dataset}
\format{
A list with 6 elements:
\describe{
  \item{probes_expr}{probes_expr, probes_expr}
  \item{probe2gene}{probe2gene, probe2gene}
  \item{genes_expr}{genes_expr, genes_expr}
  \item{group_list}{group_list, group_list}
  \item{DEG}{DEG, DEG}
  ...
}
}
\usage{
GSE95166
}
\description{
A dataset containing eSet, probes_expr, probe2gene, genes_expr, group_list, DEG
}
\keyword{datasets}
