% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmap.R
\name{idmap}
\alias{idmap}
\title{Get Probe Annotation}
\usage{
idmap(gpl = "GPL570", type = "bioc", mirror = "tencent")
}
\arguments{
\item{gpl}{GPL(GEO platform) number, eg: GPL570}

\item{type}{source of probe anntation stored, one of "pipe", "bioc", "soft", default:"pipe"}

\item{mirror}{"tencent" only for now}
}
\value{
probe annotaions
}
\description{
\code{getGPLAnno} returns probe annotations for input gpl
}
\examples{
\dontrun{
ids=idmap('GPL570')
ids=idmap('GPL570',type='soft')
ids=idmap('GPL18084',type='pipe')
}
}
