% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmap.R
\name{printGPLInfo}
\alias{printGPLInfo}
\title{Print GPL information}
\usage{
printGPLInfo(GPL = NULL)
}
\arguments{
\item{GPL}{GPL(GEO platform) number, eg: GPL570}
}
\value{
print detail information of the input GEO platform
}
\description{
Print GPL information
}
\examples{
printGPLInfo('GPL93')
}
