% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoChina.R
\name{geoChina}
\alias{geoChina}
\title{Download expression dataset by GSE id}
\usage{
geoChina(gse = "GSE2546", mirror = "tencent", destdir = getwd())
}
\arguments{
\item{gse}{input GSE id, such as GSE1009, GSE2546, gse1009.}

\item{mirror}{"tencent" only for now.}

\item{destdir}{The destination directory for data downloads.}
}
\value{
a list of ExpressionSet, which contains the  expression matrix and phenotype data
}
\description{
\code{geoChina} will download the expression matrix and phenotype data as ExpressionSet format
     from cloud in mainland China,
     it's a alternative method for getGEO function from GEOquery package.
     geoChina('gse1009') is the same as eSet=getGEO('gse1009', getGPL = F)
}
\examples{
\dontrun{
geoChina('GSE1009',destdir=tempdir())
}
}
