% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmap.R
\name{getGPLList}
\alias{getGPLList}
\title{Get all GPL list in our package
\code{getGPLList} returns all the GPL number checklist stored in package}
\usage{
getGPLList()
}
\value{
a data.frame which contains the gpl and name of array.
}
\description{
Get all GPL list in our package
\code{getGPLList} returns all the GPL number checklist stored in package
}
